fix typechain ethers-v6 generate wrong types
Unfortunately there is a bug in typechain and we get some name clashes in the generated code. There is a PR that fixes this, but until it gets merged I’ve added a patch.
More info on the name clash:
https://github.com/dethcrypto/TypeChain/pull/887

Once this PR is merged the patch could be removed.